@@ -8,7 +8,7 @@
 import Foundation
 
 /// A collection of story promos that should be displayed as a scrollable carousel.
-public struct FDCarousel: FDCollection, Codable, Equatable, Hashable {
+public struct FDCarousel: FDCollection, Codable, Equatable, Hashable, Sendable {
     /// Used for decoding `FDItem`. This is always `Carousel`.
     internal let type: String
 
