@@ -8,7 +8,7 @@
 import Foundation
 
 /// The list of ordered items to be displayed to the user for a page returned by the API.
-public enum FDItem: Codable, Equatable, Hashable {
+public enum FDItem: Codable, Equatable, Hashable, Sendable {
     case billboard(FDBillboard)
     case hierarchicalCollection(FDHierarchicalCollection)
     case collectionHeader(FDCollectionHeader)
