@@ -10,7 +10,7 @@ import Foundation
 /// The main response object from the API.
 ///
 /// This could return a list of links to stories, or the contents of a single story.
-public struct FDResult: Codable, Equatable, Hashable {
+public struct FDResult: Codable, Equatable, Hashable, Sendable {
     /// The main data from the API response.
     public var data: FDData
 
