@@ -8,7 +8,7 @@
 import Foundation
 
 /// An order for a content list.
-public enum FDContentListOrder: String, Codable, Equatable, Hashable, Unknownable {
+public enum FDContentListOrder: String, Codable, Equatable, Hashable, Sendable, Unknownable {
     case ordered = "ORDERED"
     case unordered = "UNORDERED"
 
