@@ -8,7 +8,7 @@
 import Foundation
 
 /// A type of media item.
-public enum FDMediaType: String, Codable, Equatable, Hashable, Unknownable {
+public enum FDMediaType: String, Codable, Equatable, Hashable, Sendable, Unknownable {
     case video = "VIDEO"
     case audio = "AUDIO"
 
