@@ -10,7 +10,7 @@ import OSLog
 
 extension Logger {
     /// The Unified Logging subsystem name
-    private static var subsystem = "BbcNews"
+    private static let subsystem = "BbcNews"
 
     /// Logging related to networking
     static let network = Logger(subsystem: Logger.subsystem, category: "network")
